Split select_executor.rs tests into focused modules
Refactored large select_executor.rs (650 lines) into 4 focused test modules:
- select_basic.rs (260 lines): Basic SELECT, wildcards, columns, ORDER BY
- select_where.rs (56 lines): WHERE clause filtering
- aggregates.rs (279 lines): COUNT, SUM, GROUP BY, HAVING
- select_joins.rs (109 lines): INNER JOIN tests

Fixed SqlValue sorting issue in test_group_by_with_count by using
sort_by with manual comparison instead of sort_by_key.

All 21 executor tests passing, no clippy warnings.

Closes #27
